Fix crash when a bug is moved while "move cam with bug" is enabled.
The type of `fce_param' should be `size_ptr' since it is used to store
a pointer, but it is instead an `int'. Change the type to `size_ptr',
which fixes the crash.

This generates one warning where a signed integer is being compared to
`fce_param', but in this case `fce_param' is being used as an integer,
so a cast is all that is needed (this fixes the warning).