Don't hang when 'L' is pressed while the game menu is open.
When the menu is open while a level is being played, the game hangs
when 'L' is pressed. This is because at src/komat/Berusky3d.cpp:829,
chani_zarad() is called on `p_handle[j]'. However, `p_ber->p_chap'
already contains `p_handle[j]', so calling chani_zarad() creates a
circular linked list which hangs later when we try to iterate over it.

It is the case that `p_ber->p_chap' already contains `p_handle[j]'
*every time* chain_zarad() is called at src/komat/Berusky3d.cpp:829,
so simply removing that call fixes the hang.