Enable demo recording.
Un-comment the demo recording code, some of which is
Windows-specific. Replace the Windows-specific code with portable
code. Also fix the indentation to match the surrounding code.

Also fix an instance where the return value of fread() is checked if
it is not equal to 0, whereas it should be checked if it is not equal
to 1 (that was my fault).

In src/kofola/Menu2.cpp, ddx2Init() is called with a `max_surfacu'
value of 10. 10 is not big enough, so replace that with 10000 (the
value with which it is called in ddxInit()).

When a demo file is being loaded, invert_matrix() is called such that
inv->_11 * src->_11 + inv->_12 * src->_21 + inv->_13 * src-31 == 0.
Before commit 1aa5802, this caused invert_matrix() to return NULL
while initializing the inverted matrix to all zeros. In commit
1aa5802, I had changed that behavior to trigger an assertion
instead. Revert to the old behavior (also in invert_matrix_copy() for
consistency) since this assertion is now triggered when loading a demo
file. This may be a problem with demo loading, and perhaps the
assertion should be there after all, so also add a comment which very
briefly explains the situation.

Note: the last move of a demo file is either not saved or not
loaded. Perhaps there is a reason for this, but I suspect this is not
intended behavior.