Don't trigger an assertion failure on zero-length camera movements.
A zero-length camera movement can occur when a bug is selected which
is already selected (for example, clicking on the selected bug in the
navigation bar or pressing TAB on a level with only one bug) and "move
cam with bug" is enabled.

Also, reset 'p_ber->kamnum' to zero even if '!(p_ber->kamera.aktivni &
GAME_KAMERA_ANIMACE)'. This is because for a zero-length camera
movement, 'p_ber->kamera.aktivni & GAME_KAMERA_ANIMACE' will be zero
when kam_stop() is called.