Enable Unicode support.
Since keys are stored as 'int's (not 'char's), I think this is all
that is needed to enable Unicode support.

What I really was trying to do was to make shift and caps lock work
properly. We can check if shift/caps are pressed and capitalize the
letters ourself, but this does not handle shifted keys other than
letters. With Unicode support enabled, it handles everything properly
even on my crazy keyboard layout (Programmer's Dvorak) where the
numbers are all shifted and in the wrong order!

I have not tested this with actual Unicode yet.