Implement TerminateThread() and shoot down the ADAS manager properly.
Before TerminateThread() was implemented, adas_Shot_Down_Manager()
would not make sure that the ADAS manager thread was terminated. This
means that it is possible for adas_ManagerProc() to try to use
SoundSource after it has been free()'d and set to NULL. In practice,
I've only been able to make this happen in GDB, but it's still good to
be safe.

I'm not sure what TerminateThread() is supposed to do with its exit
code argument, so I've just ignored it.