Fix slight pause between walking animations.
This was caused in commit ed7356f when I removed the -1 from
calc_endtime(). I did this because it was the simplest way to ensure
it would not return a negative number, and I didn't notice any
negative effects at the time. However, it turns out that this caused a
slight pause between bug walking animations (when a bug steps to a new
square). So re-add the -1 (but still make sure we don't return a
negative number). Also add a comment to avoid confusion in the future.