Fix bug walking animation when the camera is being moved.
In commit ed7356f, I made kam_stop() reset the camera even if
!(p_ber->kamera.aktivni & GAME_KAMERA_ANIMACE). But it turns out that
this messed up the bug walking animation when SHIFT/CTRL is
pressed. So revert that change, but now we need to make sure the
camera is reset after a zero-length camera movement. So in
kam_start(), simply reset the camera if
'p_ber->kamery[a_handle].endtime' is 0.