Make sure reflections are updated when mirror effects are enabled.
If mirror effects are disabled at the start of the level, and then
enabled later, not all reflections will be updated because
'p_ber->kamera.zmena' will be FALSE when ber_renderuj_scenu() is
called. So set 'p_ber->kamera.zmena' to TRUE when mirror effects are
re-enabled if they were disabled earlier.