Treat SDLK_PLUS as well as SDLK_EQUALS as K_PLUS.
On my keyboard layout, '+' is not shifted '='. This means that when I
want to type '+' (to zoom in) I actually have to press '='. So accept
SDLK_PLUS as K_PLUS, but still also accept SDLK_EQUALS as K_PLUS so we
don't annoy people who aren't cool enough to use programmer Dvorak
like me. ;-)