Fix reflection rendering.
When an object is off screen, it will not be rendered. So even if the
flag to render the reflection of that object is set, the reflection
still cannot be rendered because the object itself is not rendered. At
least I think that's what's happening. :-)

In any case, rendering an object when the reflection is visible even
if the object itself is not visible fixes the issue.