Don't hide an object if its reflection is not visible.
In the previous commit (4f64c78), I passed
'KONT_VIDITELNY | KONT_VIDITELNY_ZRC' to ber_render_list_vyrob_mesh()
when it is called to render the reflection. This was so that the
object would be shown if the reflection was visible even if the object
itself was not (see 4f64c78's commit message for more
details). However, this had the unfortunate effect of hiding the
object if the reflection was visible. This is because we passed both
'KONT_VIDITELNY | KONT_VIDITELNY_ZRC' to ber_render_list_vyrob_mesh(),
so *both* flags are also *removed* when the reflection is not visible.

To remedy the situation, make ber_render_list_vyrob_mesh() take *two*
flags: one to add to the visibility flag if the object (reflection) is
visible (this is where we pass both flags), and one to remove if the
object is not visible (this is where we pass only
KONT_VIDITELNY_ZRC). Also add an overloaded function with the old
interface for convenience.