Use iconv() instead of mbstowcs()/wcstombs().
Since mbstowcs() and wcstombs() behave differently depending on the
user's locale, that means that Berusky2 might not work in some locales
since the encoding of the data files are independent on
locales. iconv() is locale-independent, so use it instead.

Also remove support for CP1250 encoded files, since all text files
should be UTF-8 encoded.