Fix charset conversion error when compiled with optimization.
Since we didn't check the actual length of the string to be converted
in ConvertEncoding(), iconv() would sometimes read past the
terminating null byte or null wide char. So check the length of the
string with strnlen() or wcsnlen() (whichever is appropriate) and make
sure iconv() doesn't read past that.