Eliminate y-axis fighting in level 32 (and possibly others).
When you look straight down at a shadow in level 32, it
disappears. I'm not sure why this only happens in level 32 (maybe
others I haven't checked), but raising the shadow up a bit more along
the y axis seems to do the trick. The extra height is not noticeable,
and the shadow still appears to be on the ground.