Remove redundant increments to the height of shadows, etc.
These increments are not required because we already increment them in
kom_flek_zmen().

Also change the increment back from 0.1 to 0.001. In commit 020512e, I
changed the increment to fix level 32, but it turns out that the new
increment makes the bug selection circle noticeably higher. So revert
that change, and instead fix level 32 itself instead.