Enable save and demo file deletion.
Add a delete button to the save and demo loading dialogs in the main
screen. When the user clicks the delete button, a conformation menu
will come up and if the user confirms, the save or demo file will be
deleted.

Note that the delete button is only available from the load dialog in
the main screen. It is not available from the in-level load dialog.