Raise the bug's shadow slightly.
In level 323, when the bug walks near the edge closest to the exit,
the bug's shadow is not fully shown. Raising the bug's shadow slightly
fixes this, and, unlike for the selection circle, the extra height is
not noticeable.

And unlike in level 32, we can't just raise the 'y_start' field of the
level header, because then things start getting weird.