Don't try to guess the extension of files that cannot be loaded.
If an extension was guessed, this can cause problems when a level is
restarted, because the texture will be loaded again even if it was
already loaded. This is because the name of the texture with the
extension as it was read from the file will be matched against the
name of the texture with the guessed extension. Since these are
different, the texture will be loaded again.