Don't allow duplicate player names.
If two players have the same name, this can be very confusing for the
user, since the players will be displayed with the same name in the
menu, making it impossible to tell them apart. So if the user tries to
create a player with a name that already exists, then tell the user
that it's a duplicate and don't allow the user to create the player.